pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "pixel-cleaner"
version = "0.1.0"
description = "CLI pipeline for converting fake pixel art into clean indexed pixel art."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "numpy>=1.26",
  "Pillow>=10",
]

[project.scripts]
pixel = "pixel.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
